pyproject.toml
[project]
name = "fknc-adb-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "easyocr>=1.7.2",
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[project.scripts]
fknc-adb-helper = "fknc_adb_helper.__init__:main"

[tool.uv]
package = true

[[tool.uv.index]]
name = "bfsu"
url = "https://mirrors.bfsu.edu.cn/pypi/web/simple"
default = true

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
default = false
